refactor(x-bus): move to x-components
