--- packages/x-bus/.gitignore
@@ -1,3 +1,0 @@
-# Prettier CHANGELOG.md ignore
-/CHANGELOG.md
-/dist
