--- packages/x-bus/README.md
@@ -1,8 +1,0 @@
-# x-bus
-
-`x-bus` is a library that provides an event bus to help with event orchestration.
-
-## Contributing
-
-To start contributing to the project, please take a look at our
-[Contributing Guide](https://github.com/empathyco/x/blob/main/.github/CONTRIBUTING.md).
