--- packages/x-bus/eslint.config.mjs
@@ -1,16 +1,0 @@
-import { empathyco } from '@empathyco/eslint-config'
-
-export default empathyco(
-  {
-    rules: {
-      'ts/no-unsafe-assignment': 'off',
-      'ts/no-unsafe-return': 'off',
-    },
-  },
-  {
-    files: ['**/*.spec.{ts,tsx,js,jsx}'],
-    rules: {
-      'ts/no-floating-promises': 'off',
-    },
-  },
-)
