--- packages/x-bus/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  moduleFileExtensions: ['ts', 'js'],
-  preset: 'ts-jest',
-  testMatch: ['<rootDir>/**/*.spec.ts'],
-  testEnvironment: 'jsdom',
-}
