--- packages/x-bus/tsconfig.esm.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "esnext",
-    "declaration": false,
-    "declarationDir": null,
-    "noEmit": false,
-    "outDir": "./dist/esm"
-  },
-  "exclude": ["**/__tests__/**"]
-}
