--- packages/x-bus/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2020",
-    "lib": ["dom", "esnext"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "types": ["jest"],
-    "strict": true,
-    "declaration": true,
-    "declarationDir": "./dist/types",
-    "importHelpers": true,
-    "inlineSources": true,
-    "noEmit": true,
-    "outDir": "./dist/cjs",
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["node_modules"]
-}
