+++ packages/x-components/src/composables/use-x-bus.ts
@@ -1,8 +1,8 @@
-import type { EventPayload, SubjectPayload, XBus } from '@empathyco/x-bus'
 import type { ComponentPublicInstance, Ref } from 'vue'
 import type { FeatureLocation } from '../types/origin'
 import type { XEvent, XEventPayload, XEventsTypes } from '../wiring/events.types'
 import type { WireMetadata } from '../wiring/wiring.types'
+import type { EventPayload, SubjectPayload, XBus } from '../x-bus'
 import { Subscription } from 'rxjs'
 import { getCurrentInstance, inject, isRef, onBeforeUnmount } from 'vue'
 import { getRootXComponent, getXComponentXModuleName } from '../components/x-component.utils'
