+++ packages/x-components/src/plugins/x-emitters.ts
@@ -1,9 +1,9 @@
-import type { XBus } from '@empathyco/x-bus'
 import type { Dictionary } from '@empathyco/x-utils'
 import type { Store } from 'vuex'
 import type { AnySimpleStateSelector, AnyStateSelector } from '../store/utils/store-emitters.utils'
 import type { XEventPayload, XEventsTypes } from '../wiring/events.types'
 import type { WireMetadata } from '../wiring/wiring.types'
+import type { XBus } from '../x-bus'
 import type { AnyXModule } from '../x-modules/x-modules.types'
 import { forEach } from '@empathyco/x-utils'
 import { getGettersProxyFromModule } from '../store/utils/getters-proxy.utils'
