+++ packages/x-components/src/plugins/x-plugin.ts
@@ -1,4 +1,3 @@
-import type { EventPayload, SubjectPayload, XBus } from '@empathyco/x-bus'
 import type { XComponentsAdapter } from '@empathyco/x-types'
 import type { Dictionary } from '@empathyco/x-utils'
 import type { Observable } from 'rxjs'
@@ -7,6 +6,7 @@ import type { Module, Store } from 'vuex'
 import type { AnyXStoreModule, RootXStoreState } from '../store/store.types'
 import type { XEvent, XEventsTypes } from '../wiring/events.types'
 import type { AnyWire, WireMetadata } from '../wiring/wiring.types'
+import type { EventPayload, SubjectPayload, XBus } from '../x-bus'
 import type { AnyXModule, XModuleName } from '../x-modules/x-modules.types'
 import type { AnyXStoreModuleOption, XModuleOptions, XPluginOptions } from './x-plugin.types'
 import { deepMerge } from '@empathyco/x-deep-merge'
