+++ packages/x-components/src/wiring/__tests__/operators-testing-utils.ts
@@ -1,8 +1,8 @@
-import type { EventPayload, SubjectPayload, XBus } from '@empathyco/x-bus'
 import type { DeepPartial } from '@empathyco/x-utils'
 import type { Observable } from 'rxjs'
 import type { Store } from 'vuex'
 import type { RootXStoreState } from '../../store/index'
+import type { EventPayload, SubjectPayload, XBus } from '../../x-bus'
 import type { XEvent, XEventPayload, XEventsTypes } from '../events.types'
 import type { Wire, WireMetadata } from '../wiring.types'
 import { mount } from '@vue/test-utils'
