+++ packages/x-components/src/wiring/wires-operators.utils.ts
@@ -1,8 +1,8 @@
-import type { XBus } from '@empathyco/x-bus'
 import type { Observable } from 'rxjs'
 import type { Store } from 'vuex'
 import type { RootXStoreState } from '../store/store.types'
 import type { MaybeArray } from '../utils/types'
+import type { XBus } from '../x-bus'
 import type { XEvent, XEventsTypes } from './events.types'
 import type { TimedWireOperatorOptions, TimeSelector, WireMetadata } from './wiring.types'
 import { race, Subject, timer } from 'rxjs'
