--- packages/x-priority-queue/.gitignore
@@ -1,3 +1,0 @@
-# Prettier CHANGELOG.md ignore
-/CHANGELOG.md
-/dist
