--- packages/x-priority-queue/CHANGELOG.md
@@ -1,201 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## <small>1.0.3-alpha.5 (2025-06-26)</small>
-
-* chore(deps): bump high confidence packages (#1810) ([892f1b2](https://github.com/empathyco/x/commit/892f1b2)), closes [#1810](https://github.com/empathyco/x/issues/1810)
-
-
-
-
-
-## <small>1.0.3-alpha.4 (2025-06-06)</small>
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.3-alpha.3](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.3-alpha.2...@empathyco/x-priority-queue@1.0.3-alpha.3) (2025-04-14)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.3-alpha.2](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.3-alpha.1...@empathyco/x-priority-queue@1.0.3-alpha.2) (2025-01-22)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.3-alpha.1](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.3-alpha.0...@empathyco/x-priority-queue@1.0.3-alpha.1) (2024-02-05)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.3-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.2...@empathyco/x-priority-queue@1.0.3-alpha.0) (2023-11-29)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.2](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.2-alpha.0...@empathyco/x-priority-queue@1.0.2) (2023-11-24)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.2-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.1-alpha.0...@empathyco/x-priority-queue@1.0.2-alpha.0) (2023-11-22)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.1](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.1-alpha.0...@empathyco/x-priority-queue@1.0.1) (2023-11-21)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-## [1.0.1-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.4...@empathyco/x-priority-queue@1.0.1-alpha.0) (2023-10-24)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-
-
-
-
-# Change Log
-
-All notable changes to this project will be documented in this file. See
-[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.0.0](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.4...@empathyco/x-priority-queue@1.0.0) (2023-10-23)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [1.0.0-alpha.4](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.3...@empathyco/x-priority-queue@1.0.0-alpha.4) (2023-09-18)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [1.0.0-alpha.3](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.2...@empathyco/x-priority-queue@1.0.0-alpha.3) (2023-06-27)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [1.0.0-alpha.2](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.1...@empathyco/x-priority-queue@1.0.0-alpha.2) (2023-06-07)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [1.0.0-alpha.1](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@1.0.0-alpha.0...@empathyco/x-priority-queue@1.0.0-alpha.1) (2023-06-06)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [1.0.0-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.15...@empathyco/x-priority-queue@1.0.0-alpha.0) (2023-05-11)
-
-### ⚠ BREAKING CHANGES
-
-- **deps:** minimum node version required is 18.
-- **deps:** minimum version required is node v18
-
-### Continuous Integration
-
-- parallelize steps [(#1174)](https://github.com/empathyco/x/pull/1174)
-- restore build workflow [(#1194)](https://github.com/empathyco/x/pull/1194)
-
-### Others
-
-- **deps:** update node to 18 [(#1196)](https://github.com/empathyco/x/pull/1196)
-- **deps:** update node to v18 [(#1175)](https://github.com/empathyco/x/pull/1175)
-
-## [0.1.0-alpha.15](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.14...@empathyco/x-priority-queue@0.1.0-alpha.15) (2023-05-03)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [0.1.0-alpha.14](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.13...@empathyco/x-priority-queue@0.1.0-alpha.14) (2023-04-12)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [0.1.0-alpha.13](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.12...@empathyco/x-priority-queue@0.1.0-alpha.13) (2023-03-07)
-
-### Bug Fixes
-
-- **deps:** remove `only-allow` pnpm [(#1097)](https://github.com/empathyco/x/pull/1097)
-
-## [0.1.0-alpha.12](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.11...@empathyco/x-priority-queue@0.1.0-alpha.12) (2023-02-08)
-
-### Continuous Integration
-
-- refactor `lint` command to benefit from `nx parallelisation` [(#1051)](https://github.com/empathyco/x/pull/1051)
-
-## [0.1.0-alpha.11](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.10...@empathyco/x-priority-queue@0.1.0-alpha.11) (2023-02-07)
-
-### Continuous Integration
-
-- migrate from `npm` & `lerna bootstrap` to `pnpm` [(#1047)](https://github.com/empathyco/x/pull/1047)
-
-## [0.1.0-alpha.10](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.9...@empathyco/x-priority-queue@0.1.0-alpha.10) (2023-01-25)
-
-### Build System
-
-- **dependencies:** update `typescript` to `4.9.4` and its dependencies to their latest version
-  [(#993)](https://github.com/empathyco/x/pull/993)
-
-## [0.1.0-alpha.9](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.8...@empathyco/x-priority-queue@0.1.0-alpha.9) (2023-01-20)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [0.1.0-alpha.8](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.7...@empathyco/x-priority-queue@0.1.0-alpha.8) (2023-01-18)
-
-### Features
-
-- improve generics types [(#1012)](https://github.com/empathyco/x/pull/1012)
-
-## [0.1.0-alpha.7](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.6...@empathyco/x-priority-queue@0.1.0-alpha.7) (2023-01-12)
-
-### Features
-
-- **priority-queue:** improve types [(#992)](https://github.com/empathyco/x/pull/992)
-
-## [0.1.0-alpha.6](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.5...@empathyco/x-priority-queue@0.1.0-alpha.6) (2023-01-06)
-
-### Build System
-
-- **deps:** update dependencies
-
-## [0.1.0-alpha.5](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.4...@empathyco/x-priority-queue@0.1.0-alpha.5) (2022-12-19)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
-
-## [0.1.0-alpha.4](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.3...@empathyco/x-priority-queue@0.1.0-alpha.4) (2022-12-16)
-
-### Continuous Integration
-
-- use matrix steps and jest projects [(#919)](https://github.com/empathyco/x/pull/919)
-
-## [0.1.0-alpha.3](https://github.com/empathyco/x/compare/@empathyco/x-priority-queue@0.1.0-alpha.2...@empathyco/x-priority-queue@0.1.0-alpha.3) (2022-11-28)
-
-### Features
-
-- **priority-queue:** add `XPriorityQueue` [(#876)](https://github.com/empathyco/x/pull/876)
-
-## 0.1.0-alpha.2 (2022-11-22)
-
-**Note:** Version bump only for package @empathyco/x-priority-queue
