--- packages/x-priority-queue/README.md
@@ -1,3 +1,0 @@
-# `x-priority-queue`
-
-A priority queue implementation in typescript allowing cancellable elements.
