--- packages/x-priority-queue/eslint.config.mjs
@@ -1,3 +1,0 @@
-import { empathyco } from '@empathyco/eslint-config'
-
-export default empathyco()
