--- packages/x-priority-queue/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  moduleFileExtensions: ['ts', 'js'],
-  preset: 'ts-jest',
-  rootDir: 'src',
-  testMatch: ['<rootDir>/**/*.spec.ts'],
-}
