--- packages/x-priority-queue/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "@empathyco/x-priority-queue",
-  "version": "1.0.3-alpha.5",
-  "description": "A priority queue implementation in typescript allowing cancellable elements",
-  "author": "Empathy Systems Corporation S.L.",
-  "license": "Apache-2.0",
-  "homepage": "https://github.com/empathyco/x/tree/main/packages/x-priority-queue#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/empathyco/x.git",
-    "directory": "packages/x-priority-queue"
-  },
-  "bugs": {
-    "url": "https://github.com/empathyco/x/issues"
-  },
-  "keywords": [
-    "priority",
-    "queue",
-    "cancellable"
-  ],
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "engines": {
-    "node": ">=22"
-  },
-  "scripts": {
-    "prebuild": "rimraf dist ./*.tgz",
-    "build": "concurrently \"pnpm run build:*\"",
-    "build:cjs": "tsc --project tsconfig.cjs.json",
-    "build:esm": "tsc --project tsconfig.esm.json",
-    "postbuild": "pnpm pack",
-    "lint": "eslint --fix",
-    "lint:check": "eslint",
-    "lint:inspect": "eslint --inspect-config .",
-    "format": "prettier --write .",
-    "format:check": "prettier --check .",
-    "test": "jest",
-    "test:unit": "jest",
-    "prepublishOnly": "pnpm run build"
-  },
-  "dependencies": {
-    "@empathyco/x-utils": "^1.0.3-alpha.5",
-    "tslib": "~2.8.1"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.0",
-    "concurrently": "8.2.2",
-    "jest": "27.5.0",
-    "rimraf": "3.0.2",
-    "ts-jest": "27.1.0",
-    "typescript": "4.9.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
